Finish reorganisation and metadata
